Skip a column

Use this function to ignore a column when reading in a file. To skip all columns not otherwise specified, use cols_only().

col_skip()

See also

Other parsers: parse_datetime, parse_factor, parse_guess, parse_logical, parse_number