Guess encoding of file

Uses stringi::stri_enc_detect(): see the documentation there for caveats.

guess_encoding(file, n_max = 10000, threshold = 0.2)

Arguments

file
A character string specifying an input as specified in datasource(), a raw vector, or a list of raw vectors.

n_max
Number of lines to read. If n_max is -1, all lines in file will be read.

threshold
Only report guesses above this threshold of certainty.

Value

A tibble

Examples

guess_encoding(readr_example("mtcars.csv"))

#> [38;5;246m# A tibble: 1 x 2[39m
#>   encoding confidence
#>   [3m[38;5;246m<chr>[39m[23m         [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m1[39m ASCII             1

guess_encoding(read_lines_raw(readr_example("mtcars.csv")))

#> [38;5;246m# A tibble: 1 x 2[39m
#>   encoding confidence
#>   [3m[38;5;246m<chr>[39m[23m         [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m1[39m ASCII             1

guess_encoding(read_file_raw(readr_example("mtcars.csv")))

#> [38;5;246m# A tibble: 1 x 2[39m
#>   encoding confidence
#>   [3m[38;5;246m<chr>[39m[23m         [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m1[39m ASCII             1

guess_encoding("a\n\u00b5\u00b5")

#> [38;5;246m# A tibble: 1 x 2[39m
#>   encoding confidence
#>   [3m[38;5;246m<chr>[39m[23m         [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m1[39m UTF-8           0.8