Parse a character vector into a factor

Parse a character vector into a factor

parse_factor(x, levels, ordered = FALSE, na = c("", "NA"),
  locale = default_locale())

col_factor(levels, ordered = FALSE)

Arguments

x

Character vector of values to parse.

levels

Character vector providing set of allowed levels.

ordered

Is it an ordered factor?

na

Character vector of strings to use for missing values. Set this option to character() to indicate no missing values.

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

See also

Other parser: parse_atomic, parse_datetime, parse_guess, parse_number

Examples

parse_factor(c("a", "b"), letters)

#> [1] a b
#> Levels: a b c d e f g h i j k l m n o p q r s t u v w x y z
#>