readr: Read Rectangular Text Data

The goal of 'readr' is to provide a fast and friendly way to read rectangular data (like 'csv', 'tsv', and 'fwf'). It is designed to flexibly parse many types of data found in the wild, while still cleanly failing when data unexpectedly changes.

See also

Useful links:

http://readr.tidyverse.org

https://github.com/tidyverse/readr

Report bugs at https://github.com/tidyverse/readr/issues