readr_threads {readr}	R Documentation

Determine how many threads readr should use when processing

Description

The number of threads returned can be set by

The global option readr.num_threads

The environment variable VROOM_THREADS

The value of parallel::detectCores()

Usage

readr_threads()

[Package readr version 2.0.2 Index]