with_edition {readr}	R Documentation

Temporarily change the active readr edition

Description

with_edition() allows you to change the active edition of readr for a given block of code. local_edition() allows you to change the active edition of readr until the end of the current function or file.

Usage

with_edition(edition, code)

local_edition(edition, env = parent.frame())

Arguments

edition
Should be a single integer.

code
Code to run with the changed edition.

env
Environment that controls scope of changes. For expert use only.

[Package readr version 2.0.2 Index]