col_skip {readr}	R Documentation

Skip a column

Description

Use this function to ignore a column when reading in a file. To skip all columns not otherwise specified, use cols_only().

Usage

col_skip()

See Also

Other parsers: cols_condense(), cols(), parse_datetime(), parse_factor(), parse_guess(), parse_logical(), parse_number(), parse_vector()

[Package readr version 2.1.2 Index]