Create a source object.

Create a source object.

datasource(file, skip = 0, comment = "")

Arguments

file

Either a path to a file, a connection, or literal data (either a single string or a raw vector). Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded & decompressed. Literal data is most useful for examples and tests. It must contain at least one new line to be recognised as data (instead of a path).

skip

Number of lines to skip before reading data.

Examples

# Literal csv
datasource("a,b,c\n1,2,3")

#> [[1]]
#> [1] "a,b,c\n1,2,3"
#> 
#> $skip
#> [1] 0
#> 
#> $comment
#> [1] ""
#> 
#> attr(,"class")
#> [1] "source_string" "source"       
#> 

datasource(charToRaw("a,b,c\n1,2,3"))

#> [[1]]
#>  [1] 61 2c 62 2c 63 0a 31 2c 32 2c 33
#> 
#> $skip
#> [1] 0
#> 
#> $comment
#> [1] ""
#> 
#> attr(,"class")
#> [1] "source_raw" "source"    
#> 

# Strings
datasource(readr_example("mtcars.csv"))

#> [[1]]
#> [1] "/Users/hadley/Documents/ingest/readr/inst/extdata/mtcars.csv"
#> 
#> $skip
#> [1] 0
#> 
#> $comment
#> [1] ""
#> 
#> attr(,"class")
#> [1] "source_file" "source"     
#> 

datasource(readr_example("mtcars.csv.bz2"))

#> [[1]]
#>    [1] 22 6d 70 67 22 2c 22 63 79 6c 22 2c 22 64 69 73 70 22 2c 22 68 70 22 2c
#>   [25] 22 64 72 61 74 22 2c 22 77 74 22 2c 22 71 73 65 63 22 2c 22 76 73 22 2c
#>   [49] 22 61 6d 22 2c 22 67 65 61 72 22 2c 22 63 61 72 62 22 0a 32 31 2c 36 2c
#>   [73] 31 36 30 2c 31 31 30 2c 33 2e 39 2c 32 2e 36 32 2c 31 36 2e 34 36 2c 30
#>   [97] 2c 31 2c 34 2c 34 0a 32 31 2c 36 2c 31 36 30 2c 31 31 30 2c 33 2e 39 2c
#>  [121] 32 2e 38 37 35 2c 31 37 2e 30 32 2c 30 2c 31 2c 34 2c 34 0a 32 32 2e 38
#>  [145] 2c 34 2c 31 30 38 2c 39 33 2c 33 2e 38 35 2c 32 2e 33 32 2c 31 38 2e 36
#>  [169] 31 2c 31 2c 31 2c 34 2c 31 0a 32 31 2e 34 2c 36 2c 32 35 38 2c 31 31 30
#>  [193] 2c 33 2e 30 38 2c 33 2e 32 31 35 2c 31 39 2e 34 34 2c 31 2c 30 2c 33 2c
#>  [217] 31 0a 31 38 2e 37 2c 38 2c 33 36 30 2c 31 37 35 2c 33 2e 31 35 2c 33 2e
#>  [241] 34 34 2c 31 37 2e 30 32 2c 30 2c 30 2c 33 2c 32 0a 31 38 2e 31 2c 36 2c
#>  [265] 32 32 35 2c 31 30 35 2c 32 2e 37 36 2c 33 2e 34 36 2c 32 30 2e 32 32 2c
#>  [289] 31 2c 30 2c 33 2c 31 0a 31 34 2e 33 2c 38 2c 33 36 30 2c 32 34 35 2c 33
#>  [313] 2e 32 31 2c 33 2e 35 37 2c 31 35 2e 38 34 2c 30 2c 30 2c 33 2c 34 0a 32
#>  [337] 34 2e 34 2c 34 2c 31 34 36 2e 37 2c 36 32 2c 33 2e 36 39 2c 33 2e 31 39
#>  [361] 2c 32 30 2c 31 2c 30 2c 34 2c 32 0a 32 32 2e 38 2c 34 2c 31 34 30 2e 38
#>  [385] 2c 39 35 2c 33 2e 39 32 2c 33 2e 31 35 2c 32 32 2e 39 2c 31 2c 30 2c 34
#>  [409] 2c 32 0a 31 39 2e 32 2c 36 2c 31 36 37 2e 36 2c 31 32 33 2c 33 2e 39 32
#>  [433] 2c 33 2e 34 34 2c 31 38 2e 33 2c 31 2c 30 2c 34 2c 34 0a 31 37 2e 38 2c
#>  [457] 36 2c 31 36 37 2e 36 2c 31 32 33 2c 33 2e 39 32 2c 33 2e 34 34 2c 31 38
#>  [481] 2e 39 2c 31 2c 30 2c 34 2c 34 0a 31 36 2e 34 2c 38 2c 32 37 35 2e 38 2c
#>  [505] 31 38 30 2c 33 2e 30 37 2c 34 2e 30 37 2c 31 37 2e 34 2c 30 2c 30 2c 33
#>  [529] 2c 33 0a 31 37 2e 33 2c 38 2c 32 37 35 2e 38 2c 31 38 30 2c 33 2e 30 37
#>  [553] 2c 33 2e 37 33 2c 31 37 2e 36 2c 30 2c 30 2c 33 2c 33 0a 31 35 2e 32 2c
#>  [577] 38 2c 32 37 35 2e 38 2c 31 38 30 2c 33 2e 30 37 2c 33 2e 37 38 2c 31 38
#>  [601] 2c 30 2c 30 2c 33 2c 33 0a 31 30 2e 34 2c 38 2c 34 37 32 2c 32 30 35 2c
#>  [625] 32 2e 39 33 2c 35 2e 32 35 2c 31 37 2e 39 38 2c 30 2c 30 2c 33 2c 34 0a
#>  [649] 31 30 2e 34 2c 38 2c 34 36 30 2c 32 31 35 2c 33 2c 35 2e 34 32 34 2c 31
#>  [673] 37 2e 38 32 2c 30 2c 30 2c 33 2c 34 0a 31 34 2e 37 2c 38 2c 34 34 30 2c
#>  [697] 32 33 30 2c 33 2e 32 33 2c 35 2e 33 34 35 2c 31 37 2e 34 32 2c 30 2c 30
#>  [721] 2c 33 2c 34 0a 33 32 2e 34 2c 34 2c 37 38 2e 37 2c 36 36 2c 34 2e 30 38
#>  [745] 2c 32 2e 32 2c 31 39 2e 34 37 2c 31 2c 31 2c 34 2c 31 0a 33 30 2e 34 2c
#>  [769] 34 2c 37 35 2e 37 2c 35 32 2c 34 2e 39 33 2c 31 2e 36 31 35 2c 31 38 2e
#>  [793] 35 32 2c 31 2c 31 2c 34 2c 32 0a 33 33 2e 39 2c 34 2c 37 31 2e 31 2c 36
#>  [817] 35 2c 34 2e 32 32 2c 31 2e 38 33 35 2c 31 39 2e 39 2c 31 2c 31 2c 34 2c
#>  [841] 31 0a 32 31 2e 35 2c 34 2c 31 32 30 2e 31 2c 39 37 2c 33 2e 37 2c 32 2e
#>  [865] 34 36 35 2c 32 30 2e 30 31 2c 31 2c 30 2c 33 2c 31 0a 31 35 2e 35 2c 38
#>  [889] 2c 33 31 38 2c 31 35 30 2c 32 2e 37 36 2c 33 2e 35 32 2c 31 36 2e 38 37
#>  [913] 2c 30 2c 30 2c 33 2c 32 0a 31 35 2e 32 2c 38 2c 33 30 34 2c 31 35 30 2c
#>  [937] 33 2e 31 35 2c 33 2e 34 33 35 2c 31 37 2e 33 2c 30 2c 30 2c 33 2c 32 0a
#>  [961] 31 33 2e 33 2c 38 2c 33 35 30 2c 32 34 35 2c 33 2e 37 33 2c 33 2e 38 34
#>  [985] 2c 31 35 2e 34 31 2c 30 2c 30 2c 33 2c 34 0a 31
#>  [ reached getOption("max.print") -- omitted 303 entries ]
#> 
#> $skip
#> [1] 0
#> 
#> $comment
#> [1] ""
#> 
#> attr(,"class")
#> [1] "source_raw" "source"    
#> 

datasource(readr_example("mtcars.csv.zip"))

#> [[1]]
#>    [1] 22 6d 70 67 22 2c 22 63 79 6c 22 2c 22 64 69 73 70 22 2c 22 68 70 22 2c
#>   [25] 22 64 72 61 74 22 2c 22 77 74 22 2c 22 71 73 65 63 22 2c 22 76 73 22 2c
#>   [49] 22 61 6d 22 2c 22 67 65 61 72 22 2c 22 63 61 72 62 22 0a 32 31 2c 36 2c
#>   [73] 31 36 30 2c 31 31 30 2c 33 2e 39 2c 32 2e 36 32 2c 31 36 2e 34 36 2c 30
#>   [97] 2c 31 2c 34 2c 34 0a 32 31 2c 36 2c 31 36 30 2c 31 31 30 2c 33 2e 39 2c
#>  [121] 32 2e 38 37 35 2c 31 37 2e 30 32 2c 30 2c 31 2c 34 2c 34 0a 32 32 2e 38
#>  [145] 2c 34 2c 31 30 38 2c 39 33 2c 33 2e 38 35 2c 32 2e 33 32 2c 31 38 2e 36
#>  [169] 31 2c 31 2c 31 2c 34 2c 31 0a 32 31 2e 34 2c 36 2c 32 35 38 2c 31 31 30
#>  [193] 2c 33 2e 30 38 2c 33 2e 32 31 35 2c 31 39 2e 34 34 2c 31 2c 30 2c 33 2c
#>  [217] 31 0a 31 38 2e 37 2c 38 2c 33 36 30 2c 31 37 35 2c 33 2e 31 35 2c 33 2e
#>  [241] 34 34 2c 31 37 2e 30 32 2c 30 2c 30 2c 33 2c 32 0a 31 38 2e 31 2c 36 2c
#>  [265] 32 32 35 2c 31 30 35 2c 32 2e 37 36 2c 33 2e 34 36 2c 32 30 2e 32 32 2c
#>  [289] 31 2c 30 2c 33 2c 31 0a 31 34 2e 33 2c 38 2c 33 36 30 2c 32 34 35 2c 33
#>  [313] 2e 32 31 2c 33 2e 35 37 2c 31 35 2e 38 34 2c 30 2c 30 2c 33 2c 34 0a 32
#>  [337] 34 2e 34 2c 34 2c 31 34 36 2e 37 2c 36 32 2c 33 2e 36 39 2c 33 2e 31 39
#>  [361] 2c 32 30 2c 31 2c 30 2c 34 2c 32 0a 32 32 2e 38 2c 34 2c 31 34 30 2e 38
#>  [385] 2c 39 35 2c 33 2e 39 32 2c 33 2e 31 35 2c 32 32 2e 39 2c 31 2c 30 2c 34
#>  [409] 2c 32 0a 31 39 2e 32 2c 36 2c 31 36 37 2e 36 2c 31 32 33 2c 33 2e 39 32
#>  [433] 2c 33 2e 34 34 2c 31 38 2e 33 2c 31 2c 30 2c 34 2c 34 0a 31 37 2e 38 2c
#>  [457] 36 2c 31 36 37 2e 36 2c 31 32 33 2c 33 2e 39 32 2c 33 2e 34 34 2c 31 38
#>  [481] 2e 39 2c 31 2c 30 2c 34 2c 34 0a 31 36 2e 34 2c 38 2c 32 37 35 2e 38 2c
#>  [505] 31 38 30 2c 33 2e 30 37 2c 34 2e 30 37 2c 31 37 2e 34 2c 30 2c 30 2c 33
#>  [529] 2c 33 0a 31 37 2e 33 2c 38 2c 32 37 35 2e 38 2c 31 38 30 2c 33 2e 30 37
#>  [553] 2c 33 2e 37 33 2c 31 37 2e 36 2c 30 2c 30 2c 33 2c 33 0a 31 35 2e 32 2c
#>  [577] 38 2c 32 37 35 2e 38 2c 31 38 30 2c 33 2e 30 37 2c 33 2e 37 38 2c 31 38
#>  [601] 2c 30 2c 30 2c 33 2c 33 0a 31 30 2e 34 2c 38 2c 34 37 32 2c 32 30 35 2c
#>  [625] 32 2e 39 33 2c 35 2e 32 35 2c 31 37 2e 39 38 2c 30 2c 30 2c 33 2c 34 0a
#>  [649] 31 30 2e 34 2c 38 2c 34 36 30 2c 32 31 35 2c 33 2c 35 2e 34 32 34 2c 31
#>  [673] 37 2e 38 32 2c 30 2c 30 2c 33 2c 34 0a 31 34 2e 37 2c 38 2c 34 34 30 2c
#>  [697] 32 33 30 2c 33 2e 32 33 2c 35 2e 33 34 35 2c 31 37 2e 34 32 2c 30 2c 30
#>  [721] 2c 33 2c 34 0a 33 32 2e 34 2c 34 2c 37 38 2e 37 2c 36 36 2c 34 2e 30 38
#>  [745] 2c 32 2e 32 2c 31 39 2e 34 37 2c 31 2c 31 2c 34 2c 31 0a 33 30 2e 34 2c
#>  [769] 34 2c 37 35 2e 37 2c 35 32 2c 34 2e 39 33 2c 31 2e 36 31 35 2c 31 38 2e
#>  [793] 35 32 2c 31 2c 31 2c 34 2c 32 0a 33 33 2e 39 2c 34 2c 37 31 2e 31 2c 36
#>  [817] 35 2c 34 2e 32 32 2c 31 2e 38 33 35 2c 31 39 2e 39 2c 31 2c 31 2c 34 2c
#>  [841] 31 0a 32 31 2e 35 2c 34 2c 31 32 30 2e 31 2c 39 37 2c 33 2e 37 2c 32 2e
#>  [865] 34 36 35 2c 32 30 2e 30 31 2c 31 2c 30 2c 33 2c 31 0a 31 35 2e 35 2c 38
#>  [889] 2c 33 31 38 2c 31 35 30 2c 32 2e 37 36 2c 33 2e 35 32 2c 31 36 2e 38 37
#>  [913] 2c 30 2c 30 2c 33 2c 32 0a 31 35 2e 32 2c 38 2c 33 30 34 2c 31 35 30 2c
#>  [937] 33 2e 31 35 2c 33 2e 34 33 35 2c 31 37 2e 33 2c 30 2c 30 2c 33 2c 32 0a
#>  [961] 31 33 2e 33 2c 38 2c 33 35 30 2c 32 34 35 2c 33 2e 37 33 2c 33 2e 38 34
#>  [985] 2c 31 35 2e 34 31 2c 30 2c 30 2c 33 2c 34 0a 31
#>  [ reached getOption("max.print") -- omitted 303 entries ]
#> 
#> $skip
#> [1] 0
#> 
#> $comment
#> [1] ""
#> 
#> attr(,"class")
#> [1] "source_raw" "source"    
#> 

## Not run: ------------------------------------
# datasource("https://github.com/tidyverse/readr/raw/master/inst/extdata/mtcars.csv")
## ---------------------------------------------

# Connection
con <- rawConnection(charToRaw("abc\n123"))
datasource(con)

#> [[1]]
#> [1] 61 62 63 0a 31 32 33
#> 
#> $skip
#> [1] 0
#> 
#> $comment
#> [1] ""
#> 
#> attr(,"class")
#> [1] "source_raw" "source"    
#> 

close(con)