guess_encoding {readr}	R Documentation

Guess encoding of file

Description

Uses stringi::stri_enc_detect(): see the documentation there for caveats.

Usage

guess_encoding(file, n_max = 10000, threshold = 0.2)

Arguments

file
A character string specifying an input as specified in datasource(), a raw vector, or a list of raw vectors.

n_max
Number of lines to read. If n_max is -1, all lines in file will be read.

threshold
Only report guesses above this threshold of certainty.

Value

A tibble

Examples

guess_encoding(readr_example("mtcars.csv"))
guess_encoding(read_lines_raw(readr_example("mtcars.csv")))
guess_encoding(read_file_raw(readr_example("mtcars.csv")))

guess_encoding("a\n\u00b5\u00b5")

[Package readr version 2.0.2 Index]