Guess encoding of file.

Uses stri_enc_detect: see the documentation there for caveats.

guess_encoding(file, n_max = 10000, threshold = 0.2)

Arguments

file

A character string specifying an input as specified in datasource, a raw vector, or a list of raw vectors.

n_max

Number of lines to read. If n_max is -1, all lines in file will be read.

threshold

Only report guesses above this threshold of certainty.

Value

A tibble

Examples

guess_encoding(readr_example("mtcars.csv"))

#>   encoding confidence
#> 1    ASCII          1
#> 

guess_encoding(read_lines_raw(readr_example("mtcars.csv")))

#>   encoding confidence
#> 1    ASCII          1
#> 

guess_encoding(read_file_raw(readr_example("mtcars.csv")))

#>   encoding confidence
#> 1    ASCII          1
#> 

guess_encoding("a\n\u00b5\u00b5")

#> # A tibble: 1 × 2
#>   encoding confidence
#>      <chr>      <dbl>
#> 1    UTF-8        0.8
#>