parse_vector {readr}	R Documentation

Parse a character vector.

Description

Parse a character vector.

Usage

parse_vector(
  x,
  collector,
  na = c("", "NA"),
  locale = default_locale(),
  trim_ws = TRUE
)

Arguments

x
Character vector of elements to parse.

collector
Column specification.

na
Character vector of strings to interpret as missing values. Set this option to character() to indicate no missing values.

locale
The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

trim_ws
Should leading and trailing whitespace (ASCII spaces and tabs) be trimmed from each field before parsing it?

See Also

Other parsers: col_skip(), cols_condense(), cols(), parse_datetime(), parse_factor(), parse_guess(), parse_logical(), parse_number()

Examples

x <- c("1", "2", "3", "NA")
parse_vector(x, col_integer())
parse_vector(x, col_double())

[Package readr version 2.0.2 Index]