Parse a character vector.

Parse a character vector.

parse_vector(x, collector, na = c("", "NA"), locale = default_locale())

Arguments

x

Character vector of elements to parse.

collector

Column specification.

na

Character vector of strings to use for missing values. Set this option to character() to indicate no missing values.

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

Examples

x <- c("1", "2", "3", "NA")
parse_vector(x, col_integer())

#> [1]  1  2  3 NA
#> 

parse_vector(x, col_double())

#> [1]  1  2  3 NA
#>