Retrieve parsing problems

Readr functions will only throw an error if parsing fails in an unrecoverable way. However, there are lots of potential problems that you might want to know about - these are stored in the problems attribute of the output, which you can easily access with this function. stop_for_problems() will throw an error if there are any parsing problems: this is useful for automated scripts where you want to throw an error as soon as you encounter a problem.

problems(x)

stop_for_problems(x)

Arguments

x
An data frame (from read_*()) or a vector (from parse_*()).

Value

A data frame with one row for each problem and four columns:

row,col

Row and column of problem

expected

What readr expected to find

actual

What it actually got

Examples

x <- parse_integer(c("1X", "blah", "3"))

#> Warning: 2 parsing failures.
#> row col               expected actual
#>   1  -- no trailing characters   X   
#>   2  -- an integer               blah

problems(x)

#> [38;5;246m# A tibble: 2 x 4[39m
#>     row   col expected               actual
#>   [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<chr>[39m[23m                  [3m[38;5;246m<chr>[39m[23m 
#> [38;5;250m1[39m     1    [31mNA[39m no trailing characters X     
#> [38;5;250m2[39m     2    [31mNA[39m an integer             blah  

y <- parse_integer(c("1", "2", "3"))
problems(y)

#> [38;5;246m# tibble [0 × 4][39m
#> [38;5;246m# ... with 4 variables: row [3m[38;5;246m<int>[38;5;246m[23m, col [3m[38;5;246m<int>[38;5;246m[23m, expected [3m[38;5;246m<chr>[38;5;246m[23m, actual [3m[38;5;246m<chr>[38;5;246m[23m[39m