Read a delimited file by chunks

Read a delimited file by chunks

read_delim_chunked(file, callback, chunk_size = 10000, delim, quote = "\"",
  escape_backslash = FALSE, escape_double = TRUE, col_names = TRUE,
  col_types = NULL, locale = default_locale(), na = c("", "NA"),
  quoted_na = TRUE, comment = "", trim_ws = FALSE, skip = 0,
  guess_max = min(1000, chunk_size), progress = interactive())

read_csv_chunked(file, callback, chunk_size = 10000, col_names = TRUE,
  col_types = NULL, locale = default_locale(), na = c("", "NA"),
  quoted_na = TRUE, comment = "", trim_ws = TRUE, skip = 0,
  guess_max = min(1000, chunk_size), progress = interactive())

read_csv2_chunked(file, callback, chunk_size = 10000, col_names = TRUE,
  col_types = NULL, locale = default_locale(), na = c("", "NA"),
  quoted_na = TRUE, comment = "", trim_ws = TRUE, skip = 0,
  guess_max = min(1000, chunk_size), progress = interactive())

read_tsv_chunked(file, callback, chunk_size = 10000, col_names = TRUE,
  col_types = NULL, locale = default_locale(), na = c("", "NA"),
  quoted_na = TRUE, comment = "", trim_ws = TRUE, skip = 0,
  guess_max = min(1000, chunk_size), progress = interactive())

Arguments

file

Either a path to a file, a connection, or literal data (either a single string or a raw vector). Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded & decompressed. Literal data is most useful for examples and tests. It must contain at least one new line to be recognised as data (instead of a path).

callback

A callback function to call on each chunk

chunk_size

The number of rows to include in each chunk

delim

Single character used to separate fields within a record.

quote

Single character used to quote strings.

escape_backslash

Does the file use backslashes to escape special characters? This is more general than escape_double as backslashes can be used to escape the delimeter character, the quote characer, or to add special characters like \n.

escape_double

Does the file escape quotes by doubling them? i.e. If this option is TRUE, the value """" represents a single quote, \".

col_names

Either TRUE, FALSE or a character vector of column names. If TRUE, the first row of the input will be used as the column names, and will not be included in the data frame. If FALSE, column names will be generated automatically: X1, X2, X3 etc. If col_names is a character vector, the values will be used as the names of the columns, and the first row of the input will be read into the first row of the output data frame. Missing (NA) column names will generate a warning, and be filled in with dummy names X1, X2 etc. Duplicate column names will generate a warning and be made unique with a numeric prefix.

col_types

One of NULL, a cols specification, or a string. See vignette("column-types") for more details. If NULL, all column types will be imputed from the first 1000 rows on the input. This is convenient (and fast), but not robust. If the imputation fails, you'll need to supply the correct types yourself. If a column specification created by cols, it must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only. Alternatively, you can use a compact string representation where each character represents one column: c = character, i = integer, n = number, d = double, l = logical, D = date, T = date time, t = time, ? = guess, or _/- to skip the column.

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

na

Character vector of strings to use for missing values. Set this option to character() to indicate no missing values.

quoted_na

Should missing values inside quotes be treated as missing values (the default) or strings.

comment

A string used to identify comments. Any text after the comment characters will be silently ignored.

trim_ws

Should leading and trailing whitespace be trimmed from each field before parsing it?

skip

Number of lines to skip before reading data.

guess_max

Maximum number of records to use for guessing column types.

progress

Display a progress bar? By default it will only display in an interactive session. The display is updated every 50,000 values and will only display if estimated reading time is 5 seconds or more.

See also

Other chunked: callback, read_lines_chunked

Examples

# Cars with 3 gears
f <- function(x, pos) subset(x, gear == 3)
read_csv_chunked(readr_example("mtcars.csv"), DataFrameCallback$new(f), chunk_size = 5)

Parsed with column specification:
cols(
  mpg = col_double(),
  cyl = col_integer(),
  disp = col_integer(),
  hp = col_integer(),
  drat = col_double(),
  wt = col_double(),
  qsec = col_double(),
  vs = col_integer(),
  am = col_integer(),
  gear = col_integer(),
  carb = col_integer()
)

#> # A tibble: 15 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> *  <dbl> <int> <int> <int> <dbl> <dbl> <dbl> <int> <int> <int> <int>
#> 1   21.4     6   258   110  3.08 3.215 19.44     1     0     3     1
#> 2   18.7     8   360   175  3.15 3.440 17.02     0     0     3     2
#> 3   18.1     6   225   105  2.76 3.460 20.22     1     0     3     1
#> 4   14.3     8   360   245  3.21 3.570 15.84     0     0     3     4
#> 5   16.4     8    NA   180  3.07 4.070 17.40     0     0     3     3
#> 6   17.3     8    NA   180  3.07 3.730 17.60     0     0     3     3
#> 7   15.2     8    NA   180  3.07 3.780 18.00     0     0     3     3
#> 8   10.4     8   472   205  2.93 5.250 17.98     0     0     3     4
#> 9   10.4     8   460   215  3.00 5.424 17.82     0     0     3     4
#> 10  14.7     8   440   230  3.23 5.345 17.42     0     0     3     4
#> 11  21.5     4    NA    97  3.70 2.465 20.01     1     0     3     1
#> 12  15.5     8   318   150  2.76 3.520 16.87     0     0     3     2
#> 13  15.2     8   304   150  3.15 3.435 17.30     0     0     3     2
#> 14  13.3     8   350   245  3.73 3.840 15.41     0     0     3     4
#> 15  19.2     8   400   175  3.08 3.845 17.05     0     0     3     2
#>