Read a file into a string.

Read a file into a string.

read_file(file, locale = default_locale())

read_file_raw(file)

Arguments

file

Either a path to a file, a connection, or literal data (either a single string or a raw vector). Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded & decompressed. Literal data is most useful for examples and tests. It must contain at least one new line to be recognised as data (instead of a path).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

Value

read_file: A length 1 character vector with the entire file contents. read_lines_raw: A raw vector with the entire file contents.

Examples

read_file(file.path(R.home(), "COPYING"))

#> [1] "\t\t    GNU GENERAL PUBLIC LICENSE\n\t\t       Version 2, June 1991\n\n Copyright (C) 1989, 1991 Free Software Foundation, Inc.\n                       51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA\n Everyone is permitted to copy and distribute verbatim copies\n of this license document, but changing it is not allowed.\n\n\t\t\t    Preamble\n\n  The licenses for most software are designed to take away your\nfreedom to share and change it.  By contrast, the GNU General Public\nLicense is intended to guarantee your freedom to share and change free\nsoftware--to make sure the software is free for all its users.  This\nGeneral Public License applies to most of the Free Software\nFoundation's software and to any other program whose authors commit to\nusing it.  (Some other Free Software Foundation software is covered by\nthe GNU Library General Public License instead.)  You can apply it to\nyour programs, too.\n\n  When we speak of free software, we are referring to freedom, not\nprice.  Our General Public Licenses are designed to make sure that you\nhave the freedom to distribute copies of free software (and charge for\nthis service if you wish), that you receive source code or can get it\nif you want it, that you can change the software or use pieces of it\nin new free programs; and that you know you can do these things.\n\n  To protect your rights, we need to make restrictions that forbid\nanyone to deny you these rights or to ask you to surrender the rights.\nThese restrictions translate to certain responsibilities for you if you\ndistribute copies of the software, or if you modify it.\n\n  For example, if you distribute copies of such a program, whether\ngratis or for a fee, you must give the recipients all the rights that\nyou have.  You must make sure that they, too, receive or can get the\nsource code.  And you must show them these terms so they know their\nrights.\n\n  We protect your rights with two steps: (1) copyright the software, and\n(2) offer you this license which gives you legal permission to copy,\ndistribute and/or modify the software.\n\n  Also, for each author's protection and ours, we want to make certain\nthat everyone understands that there is no warranty for this free\nsoftware.  If the software is modified by someone else and passed on, we\nwant its recipients to know that what they have is not the original, so\nthat any problems introduced by others will not reflect on the original\nauthors' reputations.\n\n  Finally, any free program is threatened constantly by software\npatents.  We wish to avoid the danger that redistributors of a free\nprogram will individually obtain patent licenses, in effect making the\nprogram proprietary.  To prevent this, we have made it clear that any\npatent must be licensed for everyone's free use or not licensed at all.\n\n  The precise terms and conditions for copying, distribution and\nmodification follow.\n\f\n\t\t    GNU GENERAL PUBLIC LICENSE\n   TERMS AND CONDITIONS FOR COPYING, DISTRIBUTION AND MODIFICATION\n\n  0. This License applies to any program or other work which contains\na notice placed by the copyright holder saying it may be distributed\nunder the terms of this General Public License.  The \"Program\", below,\nrefers to any such program or work, and a \"work based on the Program\"\nmeans either the Program or any derivative work under copyright law:\nthat is to say, a work containing the Program or a portion of it,\neither verbatim or with modifications and/or translated into another\nlanguage.  (Hereinafter, translation is included without limitation in\nthe term \"modification\".)  Each licensee is addressed as \"you\".\n\nActivities other than copying, distribution and modification are not\ncovered by this License; they are outside its scope.  The act of\nrunning the Program is not restricted, and the output from the Program\nis covered only if its contents constitute a work based on the\nProgram (independent of having been made by running the Program).\nWhether that is true depends on what the Program does.\n\n  1. You may copy and distribute verbatim copies of the Program's\nsource code as you receive it, in any medium, provided that you\nconspicuously and appropriately publish on each copy an appropriate\ncopyright notice and disclaimer of warranty; keep intact all the\nnotices that refer to this License and to the absence of any warranty;\nand give any other recipients of the Program a copy of this License\nalong with the Program.\n\nYou may charge a fee for the physical act of transferring a copy, and\nyou may at your option offer warranty protection in exchange for a fee.\n\n  2. You may modify your copy or copies of the Program or any portion\nof it, thus forming a work based on the Program, and copy and\ndistribute such modifications or work under the terms of Section 1\nabove, provided that you also meet all of these conditions:\n\n    a) You must cause the modified files to carry prominent notices\n    stating that you changed the files and the date of any change.\n\n    b) You must cause any work that you distribute or publish, that in\n    whole or in part contains or is derived from the Program or any\n    part thereof, to be licensed as a whole at no charge to all third\n    parties under the terms of this License.\n\n    c) If the modified program normally reads commands interactively\n    when run, you must cause it, when started running for such\n    interactive use in the most ordinary way, to print or display an\n    announcement including an appropriate copyright notice and a\n    notice that there is no warranty (or else, saying that you provide\n    a warranty) and that users may redistribute the program under\n    these conditions, and telling the user how to view a copy of this\n    License.  (Exception: if the Program itself is interactive but\n    does not normally print such an announcement, your work based on\n    the Program is not required to print an announcement.)\n\f\nThese requirements apply to the modified work as a whole.  If\nidentifiable sections of that work are not derived from the Program,\nand can be reasonably considered independent and separate works in\nthemselves, then this License, and its terms, do not apply to those\nsections when you distribute them as separate works.  But when you\ndistribute the same sections as part of a whole which is a work based\non the Program, the distribution of the whole must be on the terms of\nthis License, whose permissions for other licensees extend to the\nentire whole, and thus to each and every part regardless of who wrote it.\n\nThus, it is not the intent of this section to claim rights or contest\nyour rights to work written entirely by you; rather, the intent is to\nexercise the right to control the distribution of derivative or\ncollective works based on the Program.\n\nIn addition, mere aggregation of another work not based on the Program\nwith the Program (or with a work based on the Program) on a volume of\na storage or distribution medium does not bring the other work under\nthe scope of this License.\n\n  3. You may copy and distribute the Program (or a work based on it,\nunder Section 2) in object code or executable form under the terms of\nSections 1 and 2 above provided that you also do one of the following:\n\n    a) Accompany it with the complete corresponding machine-readable\n    source code, which must be distributed under the terms of Sections\n    1 and 2 above on a medium customarily used for software interchange; or,\n\n    b) Accompany it with a written offer, valid for at least three\n    years, to give any third party, for a charge no more than your\n    cost of physically performing source distribution, a complete\n    machine-readable copy of the corresponding source code, to be\n    distributed under the terms of Sections 1 and 2 above on a medium\n    customarily used for software interchange; or,\n\n    c) Accompany it with the information you received as to the offer\n    to distribute corresponding source code.  (This alternative is\n    allowed only for noncommercial distribution and only if you\n    received the program in object code or executable form with such\n    an offer, in accord with Subsection b above.)\n\nThe source code for a work means the preferred form of the work for\nmaking modifications to it.  For an executable work, complete source\ncode means all the source code for all modules it contains, plus any\nassociated interface definition files, plus the scripts used to\ncontrol compilation and installation of the executable.  However, as a\nspecial exception, the source code distributed need not include\nanything that is normally distributed (in either source or binary\nform) with the major components (compiler, kernel, and so on) of the\noperating system on which the executable runs, unless that component\nitself accompanies the executable.\n\nIf distribution of executable or object code is made by offering\naccess to copy from a designated place, then offering equivalent\naccess to copy the source code from the same place counts as\ndistribution of the source code, even though third parties are not\ncompelled to copy the source along with the object code.\n\f\n  4. You may not copy, modify, sublicense, or distribute the Program\nexcept as expressly provided under this License.  Any attempt\notherwise to copy, modify, sublicense or distribute the Program is\nvoid, and will automatically terminate your rights under this License.\nHowever, parties who have received copies, or rights, from you under\nthis License will not have their licenses terminated so long as such\nparties remain in full compliance.\n\n  5. You are not required to accept this License, since you have not\nsigned it.  However, nothing else grants you permission to modify or\ndistribute the Program or its derivative works.  These actions are\nprohibited by law if you do not accept this License.  Therefore, by\nmodifying or distributing the Program (or any work based on the\nProgram), you indicate your acceptance of this License to do so, and\nall its terms and conditions for copying, distributing or modifying\nthe Program or works based on it.\n\n  6. Each time you redistribute the Program (or any work based on the\nProgram), the recipient automatically receives a license from the\noriginal licensor to copy, distribute or modify the Program subject to\nthese terms and conditions.  You may not impose any further\nrestrictions on the recipients' exercise of the rights granted herein.\nYou are not responsible for enforcing compliance by third parties to\nthis License.\n\n  7. If, as a consequence of a court judgment or allegation of patent\ninfringement or for any other reason (not limited to patent issues),\nconditions are imposed on you (whether by court order, agreement or\notherwise) that contradict the conditions of this License, they do not\nexcuse you from the conditions of this License.  If you cannot\ndistribute so as to satisfy simultaneously your obligations under this\nLicense and any other pertinent obligations, then as a consequence you\nmay not distribute the Program at all.  For example, if a patent\nlicense would not permit royalty-free redistribution of the Program by\nall those who receive copies directly or indirectly through you, then\nthe only way you could satisfy both it and this License would be to\nrefrain entirely from distribution of the Program.\n\nIf any portion of this section is held invalid or unenforceable under\nany particular circumstance, the balance of the section is intended to\napply and the section as a whole is intended to apply in other\ncircumstances.\n\nIt is not the purpose of this section to induce you to infringe any\npatents or other property right claims or to contest validity of any\nsuch claims; this section has the sole purpose of protecting the\nintegrity of the free software distribution system, which is\nimplemented by public license practices.  Many people have made\ngenerous contributions to the wide range of software distributed\nthrough that system in reliance on consistent application of that\nsystem; it is up to the author/donor to decide if he or she is willing\nto distribute software through any other system and a licensee cannot\nimpose that choice.\n\nThis section is intended to make thoroughly clear what is believed to\nbe a consequence of the rest of this License.\n\f\n  8. If the distribution and/or use of the Program is restricted in\ncertain countries either by patents or by copyrighted interfaces, the\noriginal copyright holder who places the Program under this License\nmay add an explicit geographical distribution limitation excluding\nthose countries, so that distribution is permitted only in or among\ncountries not thus excluded.  In such case, this License incorporates\nthe limitation as if written in the body of this License.\n\n  9. The Free Software Foundation may publish revised and/or new versions\nof the General Public License from time to time.  Such new versions will\nbe similar in spirit to the present version, but may differ in detail to\naddress new problems or concerns.\n\nEach version is given a distinguishing version number.  If the Program\nspecifies a version number of this License which applies to it and \"any\nlater version\", you have the option of following the terms and conditions\neither of that version or of any later version published by the Free\nSoftware Foundation.  If the Program does not specify a version number of\nthis License, you may choose any version ever published by the Free Software\nFoundation.\n\n  10. If you wish to incorporate parts of the Program into other free\nprograms whose distribution conditions are different, write to the author\nto ask for permission.  For software which is copyrighted by the Free\nSoftware Foundation, write to the Free Software Foundation; we sometimes\nmake exceptions for this.  Our decision will be guided by the two goals\nof preserving the free status of all derivatives of our free software and\nof promoting the sharing and reuse of software generally.\n\n\t\t\t    NO WARRANTY\n\n  11. BECAUSE THE PROGRAM IS LICENSED FREE OF CHARGE, THERE IS NO WARRANTY\nFOR THE PROGRAM, TO THE EXTENT PERMITTED BY APPLICABLE LAW.  EXCEPT WHEN\nOTHERWISE STATED IN WRITING THE COPYRIGHT HOLDERS AND/OR OTHER PARTIES\nPROVIDE THE PROGRAM \"AS IS\" WITHOUT WARRANTY OF ANY KIND, EITHER EXPRESSED\nOR IMPLIED, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF\nMERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.  THE ENTIRE RISK AS\nTO THE QUALITY AND PERFORMANCE OF THE PROGRAM IS WITH YOU.  SHOULD THE\nPROGRAM PROVE DEFECTIVE, YOU ASSUME THE COST OF ALL NECESSARY SERVICING,\nREPAIR OR CORRECTION.\n\n  12. IN NO EVENT UNLESS REQUIRED BY APPLICABLE LAW OR AGREED TO IN WRITING\nWILL ANY COPYRIGHT HOLDER, OR ANY OTHER PARTY WHO MAY MODIFY AND/OR\nREDISTRIBUTE THE PROGRAM AS PERMITTED ABOVE, BE LIABLE TO YOU FOR DAMAGES,\nINCLUDING ANY GENERAL, SPECIAL, INCIDENTAL OR CONSEQUENTIAL DAMAGES ARISING\nOUT OF THE USE OR INABILITY TO USE THE PROGRAM (INCLUDING BUT NOT LIMITED\nTO LOSS OF DATA OR DATA BEING RENDERED INACCURATE OR LOSSES SUSTAINED BY\nYOU OR THIRD PARTIES OR A FAILURE OF THE PROGRAM TO OPERATE WITH ANY OTHER\nPROGRAMS), EVEN IF SUCH HOLDER OR OTHER PARTY HAS BEEN ADVISED OF THE\nPOSSIBILITY OF SUCH DAMAGES.\n\n\t\t     END OF TERMS AND CONDITIONS\n\f\n\t    How to Apply These Terms to Your New Programs\n\n  If you develop a new program, and you want it to be of the greatest\npossible use to the public, the best way to achieve this is to make it\nfree software which everyone can redistribute and change under these terms.\n\n  To do so, attach the following notices to the program.  It is safest\nto attach them to the start of each source file to most effectively\nconvey the exclusion of warranty; and each file should have at least\nthe \"copyright\" line and a pointer to where the full notice is found.\n\n    <one line to give the program's name and a brief idea of what it does.>\n    Copyright (C) <year>  <name of author>\n\n    This program is free software; you can redistribute it and/or modify\n    it under the terms of the GNU General Public License as published by\n    the Free Software Foundation; either version 2 of the License, or\n    (at your option) any later version.\n\n    This program is distributed in the hope that it will be useful,\n    but WITHOUT ANY WARRANTY; without even the implied warranty of\n    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the\n    GNU General Public License for more details.\n\n    You should have received a copy of the GNU General Public License\n    along with this program; if not, write to the Free Software\n    Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA\n\n\nAlso add information on how to contact you by electronic and paper mail.\n\nIf the program is interactive, make it output a short notice like this\nwhen it starts in an interactive mode:\n\n    Gnomovision version 69, Copyright (C) year name of author\n    Gnomovision comes with ABSOLUTELY NO WARRANTY; for details type `show w'.\n    This is free software, and you are welcome to redistribute it\n    under certain conditions; type `show c' for details.\n\nThe hypothetical commands `show w' and `show c' should show the appropriate\nparts of the General Public License.  Of course, the commands you use may\nbe called something other than `show w' and `show c'; they could even be\nmouse-clicks or menu items--whatever suits your program.\n\nYou should also get your employer (if you work as a programmer) or your\nschool, if any, to sign a \"copyright disclaimer\" for the program, if\nnecessary.  Here is a sample; alter the names:\n\n  Yoyodyne, Inc., hereby disclaims all copyright interest in the program\n  `Gnomovision' (which makes passes at compilers) written by James Hacker.\n\n  <signature of Ty Coon>, 1 April 1989\n  Ty Coon, President of Vice\n\nThis General Public License does not permit incorporating your program into\nproprietary programs.  If your program is a subroutine library, you may\nconsider it more useful to permit linking proprietary applications with the\nlibrary.  If this is what you want to do, use the GNU Library General\nPublic License instead of this License.\n"
#> 

read_lines_raw(readr_example("mtcars.csv"))

#> [[1]]
#>  [1] 22 6d 70 67 22 2c 22 63 79 6c 22 2c 22 64 69 73 70 22 2c 22 68 70 22 2c 22
#> [26] 64 72 61 74 22 2c 22 77 74 22 2c 22 71 73 65 63 22 2c 22 76 73 22 2c 22 61
#> [51] 6d 22 2c 22 67 65 61 72 22 2c 22 63 61 72 62 22
#> 
#> [[2]]
#>  [1] 32 31 2c 36 2c 31 36 30 2c 31 31 30 2c 33 2e 39 2c 32 2e 36 32 2c 31 36 2e
#> [26] 34 36 2c 30 2c 31 2c 34 2c 34
#> 
#> [[3]]
#>  [1] 32 31 2c 36 2c 31 36 30 2c 31 31 30 2c 33 2e 39 2c 32 2e 38 37 35 2c 31 37
#> [26] 2e 30 32 2c 30 2c 31 2c 34 2c 34
#> 
#> [[4]]
#>  [1] 32 32 2e 38 2c 34 2c 31 30 38 2c 39 33 2c 33 2e 38 35 2c 32 2e 33 32 2c 31
#> [26] 38 2e 36 31 2c 31 2c 31 2c 34 2c 31
#> 
#> [[5]]
#>  [1] 32 31 2e 34 2c 36 2c 32 35 38 2c 31 31 30 2c 33 2e 30 38 2c 33 2e 32 31 35
#> [26] 2c 31 39 2e 34 34 2c 31 2c 30 2c 33 2c 31
#> 
#> [[6]]
#>  [1] 31 38 2e 37 2c 38 2c 33 36 30 2c 31 37 35 2c 33 2e 31 35 2c 33 2e 34 34 2c
#> [26] 31 37 2e 30 32 2c 30 2c 30 2c 33 2c 32
#> 
#> [[7]]
#>  [1] 31 38 2e 31 2c 36 2c 32 32 35 2c 31 30 35 2c 32 2e 37 36 2c 33 2e 34 36 2c
#> [26] 32 30 2e 32 32 2c 31 2c 30 2c 33 2c 31
#> 
#> [[8]]
#>  [1] 31 34 2e 33 2c 38 2c 33 36 30 2c 32 34 35 2c 33 2e 32 31 2c 33 2e 35 37 2c
#> [26] 31 35 2e 38 34 2c 30 2c 30 2c 33 2c 34
#> 
#> [[9]]
#>  [1] 32 34 2e 34 2c 34 2c 31 34 36 2e 37 2c 36 32 2c 33 2e 36 39 2c 33 2e 31 39
#> [26] 2c 32 30 2c 31 2c 30 2c 34 2c 32
#> 
#> [[10]]
#>  [1] 32 32 2e 38 2c 34 2c 31 34 30 2e 38 2c 39 35 2c 33 2e 39 32 2c 33 2e 31 35
#> [26] 2c 32 32 2e 39 2c 31 2c 30 2c 34 2c 32
#> 
#> [[11]]
#>  [1] 31 39 2e 32 2c 36 2c 31 36 37 2e 36 2c 31 32 33 2c 33 2e 39 32 2c 33 2e 34
#> [26] 34 2c 31 38 2e 33 2c 31 2c 30 2c 34 2c 34
#> 
#> [[12]]
#>  [1] 31 37 2e 38 2c 36 2c 31 36 37 2e 36 2c 31 32 33 2c 33 2e 39 32 2c 33 2e 34
#> [26] 34 2c 31 38 2e 39 2c 31 2c 30 2c 34 2c 34
#> 
#> [[13]]
#>  [1] 31 36 2e 34 2c 38 2c 32 37 35 2e 38 2c 31 38 30 2c 33 2e 30 37 2c 34 2e 30
#> [26] 37 2c 31 37 2e 34 2c 30 2c 30 2c 33 2c 33
#> 
#> [[14]]
#>  [1] 31 37 2e 33 2c 38 2c 32 37 35 2e 38 2c 31 38 30 2c 33 2e 30 37 2c 33 2e 37
#> [26] 33 2c 31 37 2e 36 2c 30 2c 30 2c 33 2c 33
#> 
#> [[15]]
#>  [1] 31 35 2e 32 2c 38 2c 32 37 35 2e 38 2c 31 38 30 2c 33 2e 30 37 2c 33 2e 37
#> [26] 38 2c 31 38 2c 30 2c 30 2c 33 2c 33
#> 
#> [[16]]
#>  [1] 31 30 2e 34 2c 38 2c 34 37 32 2c 32 30 35 2c 32 2e 39 33 2c 35 2e 32 35 2c
#> [26] 31 37 2e 39 38 2c 30 2c 30 2c 33 2c 34
#> 
#> [[17]]
#>  [1] 31 30 2e 34 2c 38 2c 34 36 30 2c 32 31 35 2c 33 2c 35 2e 34 32 34 2c 31 37
#> [26] 2e 38 32 2c 30 2c 30 2c 33 2c 34
#> 
#> [[18]]
#>  [1] 31 34 2e 37 2c 38 2c 34 34 30 2c 32 33 30 2c 33 2e 32 33 2c 35 2e 33 34 35
#> [26] 2c 31 37 2e 34 32 2c 30 2c 30 2c 33 2c 34
#> 
#> [[19]]
#>  [1] 33 32 2e 34 2c 34 2c 37 38 2e 37 2c 36 36 2c 34 2e 30 38 2c 32 2e 32 2c 31
#> [26] 39 2e 34 37 2c 31 2c 31 2c 34 2c 31
#> 
#> [[20]]
#>  [1] 33 30 2e 34 2c 34 2c 37 35 2e 37 2c 35 32 2c 34 2e 39 33 2c 31 2e 36 31 35
#> [26] 2c 31 38 2e 35 32 2c 31 2c 31 2c 34 2c 32
#> 
#> [[21]]
#>  [1] 33 33 2e 39 2c 34 2c 37 31 2e 31 2c 36 35 2c 34 2e 32 32 2c 31 2e 38 33 35
#> [26] 2c 31 39 2e 39 2c 31 2c 31 2c 34 2c 31
#> 
#> [[22]]
#>  [1] 32 31 2e 35 2c 34 2c 31 32 30 2e 31 2c 39 37 2c 33 2e 37 2c 32 2e 34 36 35
#> [26] 2c 32 30 2e 30 31 2c 31 2c 30 2c 33 2c 31
#> 
#> [[23]]
#>  [1] 31 35 2e 35 2c 38 2c 33 31 38 2c 31 35 30 2c 32 2e 37 36 2c 33 2e 35 32 2c
#> [26] 31 36 2e 38 37 2c 30 2c 30 2c 33 2c 32
#> 
#> [[24]]
#>  [1] 31 35 2e 32 2c 38 2c 33 30 34 2c 31 35 30 2c 33 2e 31 35 2c 33 2e 34 33 35
#> [26] 2c 31 37 2e 33 2c 30 2c 30 2c 33 2c 32
#> 
#> [[25]]
#>  [1] 31 33 2e 33 2c 38 2c 33 35 30 2c 32 34 35 2c 33 2e 37 33 2c 33 2e 38 34 2c
#> [26] 31 35 2e 34 31 2c 30 2c 30 2c 33 2c 34
#> 
#> [[26]]
#>  [1] 31 39 2e 32 2c 38 2c 34 30 30 2c 31 37 35 2c 33 2e 30 38 2c 33 2e 38 34 35
#> [26] 2c 31 37 2e 30 35 2c 30 2c 30 2c 33 2c 32
#> 
#> [[27]]
#>  [1] 32 37 2e 33 2c 34 2c 37 39 2c 36 36 2c 34 2e 30 38 2c 31 2e 39 33 35 2c 31
#> [26] 38 2e 39 2c 31 2c 31 2c 34 2c 31
#> 
#> [[28]]
#>  [1] 32 36 2c 34 2c 31 32 30 2e 33 2c 39 31 2c 34 2e 34 33 2c 32 2e 31 34 2c 31
#> [26] 36 2e 37 2c 30 2c 31 2c 35 2c 32
#> 
#> [[29]]
#>  [1] 33 30 2e 34 2c 34 2c 39 35 2e 31 2c 31 31 33 2c 33 2e 37 37 2c 31 2e 35 31
#> [26] 33 2c 31 36 2e 39 2c 31 2c 31 2c 35 2c 32
#> 
#> [[30]]
#>  [1] 31 35 2e 38 2c 38 2c 33 35 31 2c 32 36 34 2c 34 2e 32 32 2c 33 2e 31 37 2c
#> [26] 31 34 2e 35 2c 30 2c 31 2c 35 2c 34
#> 
#> [[31]]
#>  [1] 31 39 2e 37 2c 36 2c 31 34 35 2c 31 37 35 2c 33 2e 36 32 2c 32 2e 37 37 2c
#> [26] 31 35 2e 35 2c 30 2c 31 2c 35 2c 36
#> 
#> [[32]]
#>  [1] 31 35 2c 38 2c 33 30 31 2c 33 33 35 2c 33 2e 35 34 2c 33 2e 35 37 2c 31 34
#> [26] 2e 36 2c 30 2c 31 2c 35 2c 38
#> 
#> [[33]]
#>  [1] 32 31 2e 34 2c 34 2c 31 32 31 2c 31 30 39 2c 34 2e 31 31 2c 32 2e 37 38 2c
#> [26] 31 38 2e 36 2c 31 2c 31 2c 34 2c 32
#> 
#>