Read a fixed width file into a tibble

A fixed width file can be a very compact representation of numeric data. It's also very fast to parse, because every field is in the same place in every line. Unfortunately, it's painful to parse because you need to describe the length of every field. Readr aims to make it as easy as possible by providing a number of different ways to describe the field structure.

read_fwf(file, col_positions, col_types = NULL, locale = default_locale(),
  na = c("", "NA"), comment = "", skip = 0, n_max = Inf,
  guess_max = min(n_max, 1000), progress = show_progress())

fwf_empty(file, skip = 0, col_names = NULL, comment = "", n = 100L)

fwf_widths(widths, col_names = NULL)

fwf_positions(start, end = NULL, col_names = NULL)

fwf_cols(...)

Arguments

file
Either a path to a file, a connection, or literal data (either a single string or a raw vector).

Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded and decompressed.

Literal data is most useful for examples and tests. It must contain at least one new line to be recognised as data (instead of a path) or be a vector of greater than length 1.

Using a value of clipboard() will read from the system clipboard.

col_positions
Column positions, as created by fwf_empty(), fwf_widths() or fwf_positions(). To read in only selected fields, use fwf_positions(). If the width of the last column is variable (a ragged fwf file), supply the last end position as NA.

col_types
One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from the first 1000 rows on the input. This is convenient (and fast), but not robust. If the imputation fails, you'll need to supply the correct types yourself.

If a column specification created by cols(), it must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column: c = character, i = integer, n = number, d = double, l = logical, D = date, T = date time, t = time, ? = guess, or _/- to skip the column.

locale
The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

na
Character vector of strings to use for missing values. Set this option to character() to indicate no missing values.

comment
A string used to identify comments. Any text after the comment characters will be silently ignored.

skip
Number of lines to skip before reading data.

n_max
Maximum number of records to read.

guess_max
Maximum number of records to use for guessing column types.

progress
Display a progress bar? By default it will only display in an interactive session and not while knitting a document. The display is updated every 50,000 values and will only display if estimated reading time is 5 seconds or more. The automatic progress bar can be disabled by setting option readr.show_progress to FALSE.

col_names
Either NULL, or a character vector column names.

n
Number of lines the tokenizer will read to determine file structure. By default it is set to 100.

widths
Width of each field. Use NA as width of last field when reading a ragged fwf file.

start, end
Starting and ending (inclusive) positions of each field. Use NA as last end field when reading a ragged fwf file.

...
If the first element is a data frame, then it must have all numeric columns and either one or two rows. The column names are the variable names, and the column values are the variable widths if a length one vector, and variable start and end positions. Otherwise, the elements of ... are used to construct a data frame with or or two rows as above.

See also

read_table() to read fixed width files where each column is separated by whitespace.

Examples

fwf_sample <- readr_example("fwf-sample.txt")
cat(read_lines(fwf_sample))

#> John Smith          WA        418-Y11-4111 Mary Hartford       CA        319-Z19-4341 Evan Nolan          IL        219-532-c301

# You can specify column positions in several ways:
# 1. Guess based on position of empty columns
read_fwf(fwf_sample, fwf_empty(fwf_sample, col_names = c("first", "last", "state", "ssn")))

#> Parsed with column specification:
#> cols(
#>   first = col_character(),
#>   last = col_character(),
#>   state = col_character(),
#>   ssn = col_character()
#> )

#> # A tibble: 3 x 4
#>   first     last state          ssn
#>   <chr>    <chr> <chr>        <chr>
#> 1  John    Smith    WA 418-Y11-4111
#> 2  Mary Hartford    CA 319-Z19-4341
#> 3  Evan    Nolan    IL 219-532-c301

# 2. A vector of field widths
read_fwf(fwf_sample, fwf_widths(c(20, 10, 12), c("name", "state", "ssn")))

#> Parsed with column specification:
#> cols(
#>   name = col_character(),
#>   state = col_character(),
#>   ssn = col_character()
#> )

#> # A tibble: 3 x 3
#>            name state          ssn
#>           <chr> <chr>        <chr>
#> 1    John Smith    WA 418-Y11-4111
#> 2 Mary Hartford    CA 319-Z19-4341
#> 3    Evan Nolan    IL 219-532-c301

# 3. Paired vectors of start and end positions
read_fwf(fwf_sample, fwf_positions(c(1, 30), c(10, 42), c("name", "ssn")))

#> Parsed with column specification:
#> cols(
#>   name = col_character(),
#>   ssn = col_character()
#> )

#> # A tibble: 3 x 2
#>         name          ssn
#>        <chr>        <chr>
#> 1 John Smith 418-Y11-4111
#> 2 Mary Hartf 319-Z19-4341
#> 3 Evan Nolan 219-532-c301

# 4. Named arguments with start and end positions
read_fwf(fwf_sample, fwf_cols(name = c(1, 10), ssn = c(30, 42)))

#> Parsed with column specification:
#> cols(
#>   name = col_character(),
#>   ssn = col_character()
#> )

#> # A tibble: 3 x 2
#>         name          ssn
#>        <chr>        <chr>
#> 1 John Smith 418-Y11-4111
#> 2 Mary Hartf 319-Z19-4341
#> 3 Evan Nolan 219-532-c301

# 5. Named arguments with column widths
read_fwf(fwf_sample, fwf_cols(name = 20, state = 10, ssn = 12))

#> Parsed with column specification:
#> cols(
#>   name = col_character(),
#>   state = col_character(),
#>   ssn = col_character()
#> )

#> # A tibble: 3 x 3
#>            name state          ssn
#>           <chr> <chr>        <chr>
#> 1    John Smith    WA 418-Y11-4111
#> 2 Mary Hartford    CA 319-Z19-4341
#> 3    Evan Nolan    IL 219-532-c301