Read lines from a file or string.

Read lines from a file or string.

read_lines(file, skip = 0, n_max = -1L, locale = default_locale(),
  na = character(), progress = interactive())

read_lines_raw(file, skip = 0, n_max = -1L, progress = interactive())

Arguments

file

Either a path to a file, a connection, or literal data (either a single string or a raw vector). Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded & decompressed. Literal data is most useful for examples and tests. It must contain at least one new line to be recognised as data (instead of a path).

skip

Number of lines to skip before reading data.

n_max

Number of lines to read. If n_max is -1, all lines in file will be read.

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

na

Character vector of strings to use for missing values. Set this option to character() to indicate no missing values.

progress

Display a progress bar? By default it will only display in an interactive session. The display is updated every 50,000 values and will only display if estimated reading time is 5 seconds or more.

Value

read_lines: A character vector with one element for each line. read_lines_raw: A list containing a raw vector for each line.

Examples

read_lines(readr_example("mtcars.csv"))

#>  [1] "\"mpg\",\"cyl\",\"disp\",\"hp\",\"drat\",\"wt\",\"qsec\",\"vs\",\"am\",\"gear\",\"carb\""
#>  [2] "21,6,160,110,3.9,2.62,16.46,0,1,4,4"                                                     
#>  [3] "21,6,160,110,3.9,2.875,17.02,0,1,4,4"                                                    
#>  [4] "22.8,4,108,93,3.85,2.32,18.61,1,1,4,1"                                                   
#>  [5] "21.4,6,258,110,3.08,3.215,19.44,1,0,3,1"                                                 
#>  [6] "18.7,8,360,175,3.15,3.44,17.02,0,0,3,2"                                                  
#>  [7] "18.1,6,225,105,2.76,3.46,20.22,1,0,3,1"                                                  
#>  [8] "14.3,8,360,245,3.21,3.57,15.84,0,0,3,4"                                                  
#>  [9] "24.4,4,146.7,62,3.69,3.19,20,1,0,4,2"                                                    
#> [10] "22.8,4,140.8,95,3.92,3.15,22.9,1,0,4,2"                                                  
#> [11] "19.2,6,167.6,123,3.92,3.44,18.3,1,0,4,4"                                                 
#> [12] "17.8,6,167.6,123,3.92,3.44,18.9,1,0,4,4"                                                 
#> [13] "16.4,8,275.8,180,3.07,4.07,17.4,0,0,3,3"                                                 
#> [14] "17.3,8,275.8,180,3.07,3.73,17.6,0,0,3,3"                                                 
#> [15] "15.2,8,275.8,180,3.07,3.78,18,0,0,3,3"                                                   
#> [16] "10.4,8,472,205,2.93,5.25,17.98,0,0,3,4"                                                  
#> [17] "10.4,8,460,215,3,5.424,17.82,0,0,3,4"                                                    
#> [18] "14.7,8,440,230,3.23,5.345,17.42,0,0,3,4"                                                 
#> [19] "32.4,4,78.7,66,4.08,2.2,19.47,1,1,4,1"                                                   
#> [20] "30.4,4,75.7,52,4.93,1.615,18.52,1,1,4,2"                                                 
#> [21] "33.9,4,71.1,65,4.22,1.835,19.9,1,1,4,1"                                                  
#> [22] "21.5,4,120.1,97,3.7,2.465,20.01,1,0,3,1"                                                 
#> [23] "15.5,8,318,150,2.76,3.52,16.87,0,0,3,2"                                                  
#> [24] "15.2,8,304,150,3.15,3.435,17.3,0,0,3,2"                                                  
#> [25] "13.3,8,350,245,3.73,3.84,15.41,0,0,3,4"                                                  
#> [26] "19.2,8,400,175,3.08,3.845,17.05,0,0,3,2"                                                 
#> [27] "27.3,4,79,66,4.08,1.935,18.9,1,1,4,1"                                                    
#> [28] "26,4,120.3,91,4.43,2.14,16.7,0,1,5,2"                                                    
#> [29] "30.4,4,95.1,113,3.77,1.513,16.9,1,1,5,2"                                                 
#> [30] "15.8,8,351,264,4.22,3.17,14.5,0,1,5,4"                                                   
#> [31] "19.7,6,145,175,3.62,2.77,15.5,0,1,5,6"                                                   
#> [32] "15,8,301,335,3.54,3.57,14.6,0,1,5,8"                                                     
#> [33] "21.4,4,121,109,4.11,2.78,18.6,1,1,4,2"                                                   
#> 

read_lines("1\n\n2")

#> [1] "1" ""  "2"
#> 

read_lines("\n")

#> [1] ""
#> 

read_lines_raw(readr_example("mtcars.csv"))

#> [[1]]
#>  [1] 22 6d 70 67 22 2c 22 63 79 6c 22 2c 22 64 69 73 70 22 2c 22 68 70 22 2c 22
#> [26] 64 72 61 74 22 2c 22 77 74 22 2c 22 71 73 65 63 22 2c 22 76 73 22 2c 22 61
#> [51] 6d 22 2c 22 67 65 61 72 22 2c 22 63 61 72 62 22
#> 
#> [[2]]
#>  [1] 32 31 2c 36 2c 31 36 30 2c 31 31 30 2c 33 2e 39 2c 32 2e 36 32 2c 31 36 2e
#> [26] 34 36 2c 30 2c 31 2c 34 2c 34
#> 
#> [[3]]
#>  [1] 32 31 2c 36 2c 31 36 30 2c 31 31 30 2c 33 2e 39 2c 32 2e 38 37 35 2c 31 37
#> [26] 2e 30 32 2c 30 2c 31 2c 34 2c 34
#> 
#> [[4]]
#>  [1] 32 32 2e 38 2c 34 2c 31 30 38 2c 39 33 2c 33 2e 38 35 2c 32 2e 33 32 2c 31
#> [26] 38 2e 36 31 2c 31 2c 31 2c 34 2c 31
#> 
#> [[5]]
#>  [1] 32 31 2e 34 2c 36 2c 32 35 38 2c 31 31 30 2c 33 2e 30 38 2c 33 2e 32 31 35
#> [26] 2c 31 39 2e 34 34 2c 31 2c 30 2c 33 2c 31
#> 
#> [[6]]
#>  [1] 31 38 2e 37 2c 38 2c 33 36 30 2c 31 37 35 2c 33 2e 31 35 2c 33 2e 34 34 2c
#> [26] 31 37 2e 30 32 2c 30 2c 30 2c 33 2c 32
#> 
#> [[7]]
#>  [1] 31 38 2e 31 2c 36 2c 32 32 35 2c 31 30 35 2c 32 2e 37 36 2c 33 2e 34 36 2c
#> [26] 32 30 2e 32 32 2c 31 2c 30 2c 33 2c 31
#> 
#> [[8]]
#>  [1] 31 34 2e 33 2c 38 2c 33 36 30 2c 32 34 35 2c 33 2e 32 31 2c 33 2e 35 37 2c
#> [26] 31 35 2e 38 34 2c 30 2c 30 2c 33 2c 34
#> 
#> [[9]]
#>  [1] 32 34 2e 34 2c 34 2c 31 34 36 2e 37 2c 36 32 2c 33 2e 36 39 2c 33 2e 31 39
#> [26] 2c 32 30 2c 31 2c 30 2c 34 2c 32
#> 
#> [[10]]
#>  [1] 32 32 2e 38 2c 34 2c 31 34 30 2e 38 2c 39 35 2c 33 2e 39 32 2c 33 2e 31 35
#> [26] 2c 32 32 2e 39 2c 31 2c 30 2c 34 2c 32
#> 
#> [[11]]
#>  [1] 31 39 2e 32 2c 36 2c 31 36 37 2e 36 2c 31 32 33 2c 33 2e 39 32 2c 33 2e 34
#> [26] 34 2c 31 38 2e 33 2c 31 2c 30 2c 34 2c 34
#> 
#> [[12]]
#>  [1] 31 37 2e 38 2c 36 2c 31 36 37 2e 36 2c 31 32 33 2c 33 2e 39 32 2c 33 2e 34
#> [26] 34 2c 31 38 2e 39 2c 31 2c 30 2c 34 2c 34
#> 
#> [[13]]
#>  [1] 31 36 2e 34 2c 38 2c 32 37 35 2e 38 2c 31 38 30 2c 33 2e 30 37 2c 34 2e 30
#> [26] 37 2c 31 37 2e 34 2c 30 2c 30 2c 33 2c 33
#> 
#> [[14]]
#>  [1] 31 37 2e 33 2c 38 2c 32 37 35 2e 38 2c 31 38 30 2c 33 2e 30 37 2c 33 2e 37
#> [26] 33 2c 31 37 2e 36 2c 30 2c 30 2c 33 2c 33
#> 
#> [[15]]
#>  [1] 31 35 2e 32 2c 38 2c 32 37 35 2e 38 2c 31 38 30 2c 33 2e 30 37 2c 33 2e 37
#> [26] 38 2c 31 38 2c 30 2c 30 2c 33 2c 33
#> 
#> [[16]]
#>  [1] 31 30 2e 34 2c 38 2c 34 37 32 2c 32 30 35 2c 32 2e 39 33 2c 35 2e 32 35 2c
#> [26] 31 37 2e 39 38 2c 30 2c 30 2c 33 2c 34
#> 
#> [[17]]
#>  [1] 31 30 2e 34 2c 38 2c 34 36 30 2c 32 31 35 2c 33 2c 35 2e 34 32 34 2c 31 37
#> [26] 2e 38 32 2c 30 2c 30 2c 33 2c 34
#> 
#> [[18]]
#>  [1] 31 34 2e 37 2c 38 2c 34 34 30 2c 32 33 30 2c 33 2e 32 33 2c 35 2e 33 34 35
#> [26] 2c 31 37 2e 34 32 2c 30 2c 30 2c 33 2c 34
#> 
#> [[19]]
#>  [1] 33 32 2e 34 2c 34 2c 37 38 2e 37 2c 36 36 2c 34 2e 30 38 2c 32 2e 32 2c 31
#> [26] 39 2e 34 37 2c 31 2c 31 2c 34 2c 31
#> 
#> [[20]]
#>  [1] 33 30 2e 34 2c 34 2c 37 35 2e 37 2c 35 32 2c 34 2e 39 33 2c 31 2e 36 31 35
#> [26] 2c 31 38 2e 35 32 2c 31 2c 31 2c 34 2c 32
#> 
#> [[21]]
#>  [1] 33 33 2e 39 2c 34 2c 37 31 2e 31 2c 36 35 2c 34 2e 32 32 2c 31 2e 38 33 35
#> [26] 2c 31 39 2e 39 2c 31 2c 31 2c 34 2c 31
#> 
#> [[22]]
#>  [1] 32 31 2e 35 2c 34 2c 31 32 30 2e 31 2c 39 37 2c 33 2e 37 2c 32 2e 34 36 35
#> [26] 2c 32 30 2e 30 31 2c 31 2c 30 2c 33 2c 31
#> 
#> [[23]]
#>  [1] 31 35 2e 35 2c 38 2c 33 31 38 2c 31 35 30 2c 32 2e 37 36 2c 33 2e 35 32 2c
#> [26] 31 36 2e 38 37 2c 30 2c 30 2c 33 2c 32
#> 
#> [[24]]
#>  [1] 31 35 2e 32 2c 38 2c 33 30 34 2c 31 35 30 2c 33 2e 31 35 2c 33 2e 34 33 35
#> [26] 2c 31 37 2e 33 2c 30 2c 30 2c 33 2c 32
#> 
#> [[25]]
#>  [1] 31 33 2e 33 2c 38 2c 33 35 30 2c 32 34 35 2c 33 2e 37 33 2c 33 2e 38 34 2c
#> [26] 31 35 2e 34 31 2c 30 2c 30 2c 33 2c 34
#> 
#> [[26]]
#>  [1] 31 39 2e 32 2c 38 2c 34 30 30 2c 31 37 35 2c 33 2e 30 38 2c 33 2e 38 34 35
#> [26] 2c 31 37 2e 30 35 2c 30 2c 30 2c 33 2c 32
#> 
#> [[27]]
#>  [1] 32 37 2e 33 2c 34 2c 37 39 2c 36 36 2c 34 2e 30 38 2c 31 2e 39 33 35 2c 31
#> [26] 38 2e 39 2c 31 2c 31 2c 34 2c 31
#> 
#> [[28]]
#>  [1] 32 36 2c 34 2c 31 32 30 2e 33 2c 39 31 2c 34 2e 34 33 2c 32 2e 31 34 2c 31
#> [26] 36 2e 37 2c 30 2c 31 2c 35 2c 32
#> 
#> [[29]]
#>  [1] 33 30 2e 34 2c 34 2c 39 35 2e 31 2c 31 31 33 2c 33 2e 37 37 2c 31 2e 35 31
#> [26] 33 2c 31 36 2e 39 2c 31 2c 31 2c 35 2c 32
#> 
#> [[30]]
#>  [1] 31 35 2e 38 2c 38 2c 33 35 31 2c 32 36 34 2c 34 2e 32 32 2c 33 2e 31 37 2c
#> [26] 31 34 2e 35 2c 30 2c 31 2c 35 2c 34
#> 
#> [[31]]
#>  [1] 31 39 2e 37 2c 36 2c 31 34 35 2c 31 37 35 2c 33 2e 36 32 2c 32 2e 37 37 2c
#> [26] 31 35 2e 35 2c 30 2c 31 2c 35 2c 36
#> 
#> [[32]]
#>  [1] 31 35 2c 38 2c 33 30 31 2c 33 33 35 2c 33 2e 35 34 2c 33 2e 35 37 2c 31 34
#> [26] 2e 36 2c 30 2c 31 2c 35 2c 38
#> 
#> [[33]]
#>  [1] 32 31 2e 34 2c 34 2c 31 32 31 2c 31 30 39 2c 34 2e 31 31 2c 32 2e 37 38 2c
#> [26] 31 38 2e 36 2c 31 2c 31 2c 34 2c 32
#> 
#>