Read lines from a file

Reads up to n_max lines from a file. New lines are not included in the output. read_lines_raw is useful for handling data with unknown encoding.

read_lines(file, skip = 0, n_max = -1L, locale = default_locale(),
  na = character(), progress = interactive())

read_lines_raw(file, skip = 0, n_max = -1L, progress = interactive())

Arguments

file

Either a path to a file, a connection, or literal data (either a single string or a raw vector). Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded & decompressed. Literal data is most useful for examples and tests. It must contain at least one new line to be recognised as data (instead of a path).

skip

Number of lines to skip before reading data.

n_max

Number of lines to read. If n_max is -1, all lines in file will be read.

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

na

Character vector of strings to use for missing values. Set this option to character() to indicate no missing values.

progress

Display a progress bar? By default it will only display in an interactive session. The display is updated every 50,000 values and will only display if estimated reading time is 5 seconds or more.

Value

read_lines: A character vector with one element for each line. read_lines_raw: A list containing a raw vector for each line.

Examples

read_lines(file.path(R.home("doc"), "AUTHORS"), n_max = 10)

#>  [1] "Authors of R."                                                                     
#>  [2] ""                                                                                  
#>  [3] "R was initially written by Robert Gentleman and Ross Ihaka—also known as \"R & R\""
#>  [4] "of the Statistics Department of the University of Auckland."                       
#>  [5] ""                                                                                  
#>  [6] "Since mid-1997 there has been a core group with write access to the R"             
#>  [7] "source, currently consisting of"                                                   
#>  [8] ""                                                                                  
#>  [9] "Douglas Bates"                                                                     
#> [10] "John Chambers"                                                                     
#> 

read_lines_raw(file.path(R.home("doc"), "AUTHORS"), n_max = 10)

#> [[1]]
#>  [1] 41 75 74 68 6f 72 73 20 6f 66 20 52 2e
#> 
#> [[2]]
#> raw(0)
#> 
#> [[3]]
#>  [1] 52 20 77 61 73 20 69 6e 69 74 69 61 6c 6c 79 20 77 72 69 74 74 65 6e 20 62
#> [26] 79 20 52 6f 62 65 72 74 20 47 65 6e 74 6c 65 6d 61 6e 20 61 6e 64 20 52 6f
#> [51] 73 73 20 49 68 61 6b 61 e2 80 94 61 6c 73 6f 20 6b 6e 6f 77 6e 20 61 73 20
#> [76] 22 52 20 26 20 52 22
#> 
#> [[4]]
#>  [1] 6f 66 20 74 68 65 20 53 74 61 74 69 73 74 69 63 73 20 44 65 70 61 72 74 6d
#> [26] 65 6e 74 20 6f 66 20 74 68 65 20 55 6e 69 76 65 72 73 69 74 79 20 6f 66 20
#> [51] 41 75 63 6b 6c 61 6e 64 2e
#> 
#> [[5]]
#> raw(0)
#> 
#> [[6]]
#>  [1] 53 69 6e 63 65 20 6d 69 64 2d 31 39 39 37 20 74 68 65 72 65 20 68 61 73 20
#> [26] 62 65 65 6e 20 61 20 63 6f 72 65 20 67 72 6f 75 70 20 77 69 74 68 20 77 72
#> [51] 69 74 65 20 61 63 63 65 73 73 20 74 6f 20 74 68 65 20 52
#> 
#> [[7]]
#>  [1] 73 6f 75 72 63 65 2c 20 63 75 72 72 65 6e 74 6c 79 20 63 6f 6e 73 69 73 74
#> [26] 69 6e 67 20 6f 66
#> 
#> [[8]]
#> raw(0)
#> 
#> [[9]]
#>  [1] 44 6f 75 67 6c 61 73 20 42 61 74 65 73
#> 
#> [[10]]
#>  [1] 4a 6f 68 6e 20 43 68 61 6d 62 65 72 73
#> 
#>