Read whitespace-separated columns into a tibble

read_table() is designed to read the type of textual data where each column is separated by one (or more) columns of space.

read_table() is like read.table(), it allows any number of whitespace characters between columns, and the lines can be of different lengths.

spec_table() returns the column specifications rather than a data frame.

Usage

read_table(
  file,
  col_names = TRUE,
  col_types = NULL,
  locale = default_locale(),
  na = "NA",
  skip = 0,
  n_max = Inf,
  guess_max = min(n_max, 1000),
  progress = show_progress(),
  comment = "",
  show_col_types = should_show_types(),
  skip_empty_rows = TRUE
)

Arguments

file

Either a path to a file, a connection, or literal data (either a single string or a raw vector).

Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded and decompressed.

Literal data is most useful for examples and tests. To be recognised as literal data, the input must be either wrapped with I(), be a string containing at least one new line, or be a vector containing at least one string with a new line.

Using a value of clipboard() will read from the system clipboard.

col_names

Either TRUE, FALSE or a character vector of column names.

If TRUE, the first row of the input will be used as the column names, and will not be included in the data frame. If FALSE, column names will be generated automatically: X1, X2, X3 etc.

If col_names is a character vector, the values will be used as the names of the columns, and the first row of the input will be read into the first row of the output data frame.

Missing (NA) column names will generate a warning, and be filled in with dummy names ...1, ...2 etc. Duplicate column names will generate a warning and be made unique, see name_repair to control how this is done.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

na

Character vector of strings to interpret as missing values. Set this option to character() to indicate no missing values.

skip

Number of lines to skip before reading data.

n_max

Maximum number of lines to read.

guess_max

Maximum number of lines to use for guessing column types. See vignette("column-types") for more details.

progress

Display a progress bar? By default it will only display in an interactive session and not while knitting a document. The automatic progress bar can be disabled by setting option readr.show_progress to FALSE.

comment

A string used to identify comments. Any text after the comment characters will be silently ignored.

show_col_types

If FALSE, do not show the guessed column types. If TRUE always show the column types, even if they are supplied. If NULL (the default) only show the column types if they are not explicitly supplied by the col_types argument.

skip_empty_rows

Should blank rows be ignored altogether? i.e. If this option is TRUE then blank rows will not be represented at all. If it is FALSE then they will be represented by NA values in all the columns.

See also

Examples

# One corner from http://www.masseyratings.com/cf/compare.htm
massey <- readr_example("massey-rating.txt")
cat(read_file(massey))
#> UCC PAY LAZ KPK  RT   COF BIH DII ENG ACU Rank Team            Conf
#>   1   1   1   1   1     1   1   1   1   1    1 Ohio St          B10 
#>   2   2   2   2   2     2   2   2   4   2    2 Oregon           P12 
#>   3   4   3   4   3     4   3   4   2   3    3 Alabama          SEC 
#>   4   3   4   3   4     3   5   3   3   4    4 TCU              B12 
#>   6   6   6   5   5     7   6   5   6  11    5 Michigan St      B10 
#>   7   7   7   6   7     6  11   8   7   8    6 Georgia          SEC 
#>   5   5   5   7   6     8   4   6   5   5    7 Florida St       ACC 
#>   8   8   9   9  10     5   7   7  10   7    8 Baylor           B12 
#>   9  11   8  13  11    11  12   9  14   9    9 Georgia Tech     ACC 
#>  13  10  13  11   8     9  10  11   9  10   10 Mississippi      SEC 
read_table(massey)
#> 
#> ── Column specification ──────────────────────────────────────────────────
#> cols(
#>   UCC = col_double(),
#>   PAY = col_double(),
#>   LAZ = col_double(),
#>   KPK = col_double(),
#>   RT = col_double(),
#>   COF = col_double(),
#>   BIH = col_double(),
#>   DII = col_double(),
#>   ENG = col_double(),
#>   ACU = col_double(),
#>   Rank = col_double(),
#>   Team = col_character(),
#>   Conf = col_character()
#> )
#> Warning: 10 parsing failures.
#> row col   expected     actual                                                              file
#>   1  -- 13 columns 15 columns '/home/runner/work/_temp/Library/readr/extdata/massey-rating.txt'
#>   2  -- 13 columns 14 columns '/home/runner/work/_temp/Library/readr/extdata/massey-rating.txt'
#>   3  -- 13 columns 14 columns '/home/runner/work/_temp/Library/readr/extdata/massey-rating.txt'
#>   4  -- 13 columns 14 columns '/home/runner/work/_temp/Library/readr/extdata/massey-rating.txt'
#>   5  -- 13 columns 15 columns '/home/runner/work/_temp/Library/readr/extdata/massey-rating.txt'
#> ... ... .......... .......... .................................................................
#> See problems(...) for more details.
#> # A tibble: 10 × 13
#>      UCC   PAY   LAZ   KPK    RT   COF   BIH   DII   ENG   ACU  Rank Team 
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <chr>
#>  1     1     1     1     1     1     1     1     1     1     1     1 Ohio 
#>  2     2     2     2     2     2     2     2     2     4     2     2 Oreg…
#>  3     3     4     3     4     3     4     3     4     2     3     3 Alab…
#>  4     4     3     4     3     4     3     5     3     3     4     4 TCU  
#>  5     6     6     6     5     5     7     6     5     6    11     5 Mich…
#>  6     7     7     7     6     7     6    11     8     7     8     6 Geor…
#>  7     5     5     5     7     6     8     4     6     5     5     7 Flor…
#>  8     8     8     9     9    10     5     7     7    10     7     8 Bayl…
#>  9     9    11     8    13    11    11    12     9    14     9     9 Geor…
#> 10    13    10    13    11     8     9    10    11     9    10    10 Miss…
#> # … with 1 more variable: Conf <chr>

# Sample of 1978 fuel economy data from
# http://www.fueleconomy.gov/feg/epadata/78data.zip
epa <- readr_example("epa78.txt")
cat(read_file(epa))
#> ALFA ROMEO                                                                     ALFA ROMEO           78010003
#> ALFETTA                              03  81  8    74  7   89  9                ALFETTA              78010053
#> SPIDER 2000                          01                                        SPIDER 2000          78010103
#> AMC                                                                            AMC                  78020002
#> GREMLIN                              03  79  9                    79  9        GREMLIN              78020053
#> PACER                                04  89 11                    89 11        PACER                78020103
#> PACER WAGON                          07  90 26    91 26                        PACER WAGON          78020153
#> CONCORD                              04  88 12    90 11   90 11   83 16        CONCORD              78020203
#> CONCORD WAGON                        07  91 30            91 30                CONCORD WAGON        78020253
#> MATADOR COUPE                        05  97 14    97 14                        MATADOR COUPE        78020303
#> MATADOR SEDAN                        06 110 20           110 20                MATADOR SEDAN        78020353
#> MATADOR WAGON                        09 112 50           112 50                MATADOR WAGON        78020403
#> ASTON MARTIN                                                                   ASTON MARTIN         78040002
#> ASTON MARTIN                                                                   ASTON MARTIN         78040053
#> AUDI                                                                           AUDI                 78050002
#> FOX                                  03  84 11    84 11   84 11                FOX                  78050053
#> FOX WAGON                            07  83 40            83 40                FOX WAGON            78050103
#> 5000                                 04  90 15            90 15                5000                 78050153
#> AVANTI                                                                         AVANTI               78065002
#> AVANTI II                            02  75  8    75  8                        AVANTI II            78065053
read_table(epa, col_names = FALSE)
#> 
#> ── Column specification ──────────────────────────────────────────────────
#> cols(
#>   X1 = col_character(),
#>   X2 = col_character(),
#>   X3 = col_character(),
#>   X4 = col_character(),
#>   X5 = col_double()
#> )
#> Warning: 17 parsing failures.
#> row col  expected     actual                                                      file
#>   2  -- 5 columns 10 columns '/home/runner/work/_temp/Library/readr/extdata/epa78.txt'
#>   3  -- 5 columns 6 columns  '/home/runner/work/_temp/Library/readr/extdata/epa78.txt'
#>   4  -- 5 columns 3 columns  '/home/runner/work/_temp/Library/readr/extdata/epa78.txt'
#>   5  -- 5 columns 8 columns  '/home/runner/work/_temp/Library/readr/extdata/epa78.txt'
#>   6  -- 5 columns 8 columns  '/home/runner/work/_temp/Library/readr/extdata/epa78.txt'
#> ... ... ......... .......... .........................................................
#> See problems(...) for more details.
#> # A tibble: 20 × 5
#>    X1      X2     X3       X4           X5
#>    <chr>   <chr>  <chr>    <chr>     <dbl>
#>  1 ALFA    ROMEO  ALFA     ROMEO  78010003
#>  2 ALFETTA 03     81       8            74
#>  3 SPIDER  2000   01       SPIDER     2000
#>  4 AMC     AMC    78020002 NA           NA
#>  5 GREMLIN 03     79       9            79
#>  6 PACER   04     89       11           89
#>  7 PACER   WAGON  07       90           26
#>  8 CONCORD 04     88       12           90
#>  9 CONCORD WAGON  07       91           30
#> 10 MATADOR COUPE  05       97           14
#> 11 MATADOR SEDAN  06       110          20
#> 12 MATADOR WAGON  09       112          50
#> 13 ASTON   MARTIN ASTON    MARTIN 78040002
#> 14 ASTON   MARTIN ASTON    MARTIN 78040053
#> 15 AUDI    AUDI   78050002 NA           NA
#> 16 FOX     03     84       11           84
#> 17 FOX     WAGON  07       83           40
#> 18 5000    04     90       15           90
#> 19 AVANTI  AVANTI 78065002 NA           NA
#> 20 AVANTI  II     02       75            8