read_table2 {readr}	R Documentation

Read whitespace-separated columns into a tibble

Description

This function is deprecated because we renamed it to read_table() and removed the old read_table function, which was too strict for most cases and was analogous to just using read_fwf().

Usage

read_table2(
  file,
  col_names = TRUE,
  col_types = NULL,
  locale = default_locale(),
  na = "NA",
  skip = 0,
  n_max = Inf,
  guess_max = min(n_max, 1000),
  progress = show_progress(),
  comment = "",
  skip_empty_rows = TRUE
)

[Package readr version 2.0.2 Index]