readr-package {readr}	R Documentation

readr: Read Rectangular Text Data

Description

The goal of 'readr' is to provide a fast and friendly way to read rectangular data (like 'csv', 'tsv', and 'fwf'). It is designed to flexibly parse many types of data found in the wild, while still cleanly failing when data unexpectedly changes.

Author(s)

Maintainer: Jim Hester jim.hester@rstudio.com

Authors:

Hadley Wickham hadley@rstudio.com

Other contributors:

Romain Francois [contributor]

RStudio [copyright holder, funder]

https://github.com/mandreyel/ (mio library) [copyright holder]

Jukka Jylänki (grisu3 implementation) [contributor, copyright holder]

Mikkel Jørgensen (grisu3 implementation) [contributor, copyright holder]

See Also

Useful links:

https://readr.tidyverse.org

https://github.com/tidyverse/readr

Report bugs at https://github.com/tidyverse/readr/issues

[Package readr version 2.0.2 Index]