Re-convert character columns in existing data frame.

This is useful if you need to do some manual munging - you can read the columns in as character, clean it up with (e.g.) regular expressions and then let readr take another stab at parsing it.

type_convert(df, col_types = NULL, na = c("", "NA"), trim_ws = TRUE,
  locale = default_locale())

Arguments

df

A data frame.

col_types

One of NULL, a cols specification, or a string. See vignette("column-types") for more details. If NULL, all column types will be imputed from the first 1000 rows on the input. This is convenient (and fast), but not robust. If the imputation fails, you'll need to supply the correct types yourself. If a column specification created by cols, it must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only. Unlike other functions type_convert does not allow character specificatinos of col_types.

na

Character vector of strings to use for missing values. Set this option to character() to indicate no missing values.

trim_ws

Should leading and trailing whitespace be trimmed from each field before parsing it?

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

Examples

df <- data.frame(
  x = as.character(runif(10)),
  y = as.character(sample(10)),
  stringsAsFactors = FALSE
)
str(df)

#> 'data.frame':	10 obs. of  2 variables:
#>  $ x: chr  "0.0807501375675201" "0.834333037259057" "0.600760886212811" "0.157208441523835" ...
#>  $ y: chr  "9" "2" "1" "3" ...
#> 

str(type_convert(df))

Parsed with column specification:
cols(
  x = col_double(),
  y = col_integer()
)

#> 'data.frame':	10 obs. of  2 variables:
#>  $ x: num  0.0808 0.8343 0.6008 0.1572 0.0074 ...
#>  $ y: int  9 2 1 3 7 8 10 5 6 4
#> 

df <- data.frame(x = c("NA", "10"), stringsAsFactors = FALSE)
str(type_convert(df))

Parsed with column specification:
cols(
  x = col_integer()
)

#> 'data.frame':	2 obs. of  1 variable:
#>  $ x: int  NA 10
#>