Write lines/ a file

write_lines takes a character vector, appending a new line after entry. write_file takes a single string, or a raw vector, and writes it exactly as is.

write_lines(x, path, na = "NA", append = FALSE)

write_file(x, path, append = FALSE)

Arguments

x

A data frame to write to disk

path

Path to write to.

na

String used for missing values. Defaults to NA. Missing values will never be quoted; strings with the same value as na will always be quoted.

append

If FALSE, will overwrite existing file. If TRUE, will append to existing file. In both cases, if file does not exist a new file is created.

Value

The input x, invisibly.

Examples

tmp <- tempfile()

write_lines(rownames(mtcars), tmp)
read_lines(tmp)

#>  [1] "Mazda RX4"           "Mazda RX4 Wag"       "Datsun 710"         
#>  [4] "Hornet 4 Drive"      "Hornet Sportabout"   "Valiant"            
#>  [7] "Duster 360"          "Merc 240D"           "Merc 230"           
#> [10] "Merc 280"            "Merc 280C"           "Merc 450SE"         
#> [13] "Merc 450SL"          "Merc 450SLC"         "Cadillac Fleetwood" 
#> [16] "Lincoln Continental" "Chrysler Imperial"   "Fiat 128"           
#> [19] "Honda Civic"         "Toyota Corolla"      "Toyota Corona"      
#> [22] "Dodge Challenger"    "AMC Javelin"         "Camaro Z28"         
#> [25] "Pontiac Firebird"    "Fiat X1-9"           "Porsche 914-2"      
#> [28] "Lotus Europa"        "Ford Pantera L"      "Ferrari Dino"       
#> [31] "Maserati Bora"       "Volvo 142E"         
#> 

read_file(tmp) # note trailing \n

#> [1] "Mazda RX4\nMazda RX4 Wag\nDatsun 710\nHornet 4 Drive\nHornet Sportabout\nValiant\nDuster 360\nMerc 240D\nMerc 230\nMerc 280\nMerc 280C\nMerc 450SE\nMerc 450SL\nMerc 450SLC\nCadillac Fleetwood\nLincoln Continental\nChrysler Imperial\nFiat 128\nHonda Civic\nToyota Corolla\nToyota Corona\nDodge Challenger\nAMC Javelin\nCamaro Z28\nPontiac Firebird\nFiat X1-9\nPorsche 914-2\nLotus Europa\nFord Pantera L\nFerrari Dino\nMaserati Bora\nVolvo 142E\n"
#> 

write_lines(airquality$Ozone, tmp, na = "-1")
read_lines(tmp)

#>   [1] "41"  "36"  "12"  "18"  "-1"  "28"  "23"  "19"  "8"   "-1"  "7"   "16" 
#>  [13] "11"  "14"  "18"  "14"  "34"  "6"   "30"  "11"  "1"   "11"  "4"   "32" 
#>  [25] "-1"  "-1"  "-1"  "23"  "45"  "115" "37"  "-1"  "-1"  "-1"  "-1"  "-1" 
#>  [37] "-1"  "29"  "-1"  "71"  "39"  "-1"  "-1"  "23"  "-1"  "-1"  "21"  "37" 
#>  [49] "20"  "12"  "13"  "-1"  "-1"  "-1"  "-1"  "-1"  "-1"  "-1"  "-1"  "-1" 
#>  [61] "-1"  "135" "49"  "32"  "-1"  "64"  "40"  "77"  "97"  "97"  "85"  "-1" 
#>  [73] "10"  "27"  "-1"  "7"   "48"  "35"  "61"  "79"  "63"  "16"  "-1"  "-1" 
#>  [85] "80"  "108" "20"  "52"  "82"  "50"  "64"  "59"  "39"  "9"   "16"  "78" 
#>  [97] "35"  "66"  "122" "89"  "110" "-1"  "-1"  "44"  "28"  "65"  "-1"  "22" 
#> [109] "59"  "23"  "31"  "44"  "21"  "9"   "-1"  "45"  "168" "73"  "-1"  "76" 
#> [121] "118" "84"  "85"  "96"  "78"  "73"  "91"  "47"  "32"  "20"  "23"  "21" 
#> [133] "24"  "44"  "21"  "28"  "9"   "13"  "46"  "18"  "13"  "24"  "16"  "13" 
#> [145] "23"  "36"  "7"   "14"  "30"  "-1"  "14"  "18"  "20" 
#>